Move LocalFeedLoader and Feedstore collaborator to its own file in production
