+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/LocalFeedLoader.swift
@@ -1,0 +1,45 @@
+//
+//  LocalFeedLoader.swift
+//  PokemonEncyclopedia
+//
+//  Created by Ankit on 07/07/21.
+//
+
+import Foundation
+
+public class LocalFeedLoader {
+	private let store:FeedStore
+	private let currentDate: ()-> Date
+	
+	public init(store: FeedStore, currentDate: @escaping ()->Date){
+		self.store = store
+		self.currentDate = currentDate
+	}
+	
+	public func save(_ items:[FeedItem], completion: @escaping (Error?)->Void ){
+		store.deleteCachedFeed { [weak self] error in
+			guard let self = self else { return }
+			if let cacheDeletionError = error {
+				completion(cacheDeletionError)
+				return
+			}
+			self.cache(items, with: completion)
+		}
+	}
+	
+	private func cache(_ items: [FeedItem], with completion: @escaping (Error?)->Void ) {
+		store.insert(items, timestamp: currentDate()) { [weak self] error in
+			guard self != nil else {return}
+			
+			completion(error)
+		}
+	}
+}
+
+public protocol FeedStore {
+	typealias DeletionCompletion = (Error?)->Void
+	typealias InsertionCompletion = (Error?)->Void
+
+	func deleteCachedFeed(completion: @escaping DeletionCompletion)
+	func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
