+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/CacheFeedUseCaseTests.swift
@@ -8,43 +8,6 @@
 import XCTest
 import PokemonEncyclopedia
 
-class LocalFeedLoader {
-	private let store:FeedStore
-	private let currentDate: ()-> Date
-	
-	init(store: FeedStore, currentDate: @escaping ()->Date){
-		self.store = store
-		self.currentDate = currentDate
-	}
-	
-	func save(_ items:[FeedItem], completion: @escaping (Error?)->Void ){
-		store.deleteCachedFeed { [weak self] error in
-			guard let self = self else { return }
-			if let cacheDeletionError = error {
-				completion(cacheDeletionError)
-				return
-			}
-			self.cache(items, with: completion)
-		}
-	}
-	
-	private func cache(_ items: [FeedItem], with completion: @escaping (Error?)->Void ) {
-		store.insert(items, timestamp: currentDate()) { [weak self] error in
-			guard self != nil else {return}
-			
-			completion(error)
-		}
-	}
-}
-
-protocol FeedStore {
-	typealias DeletionCompletion = (Error?)->Void
-	typealias InsertionCompletion = (Error?)->Void
-
-	func deleteCachedFeed(completion: @escaping DeletionCompletion)
-	func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
-
 
 class CacheFeedUseCaseTests: XCTestCase {
